updated bash scripts to new folder/file names

+++ examples/clear.sh
@@ -3,7 +3,9 @@
 set -e
 
 for folder in */ ; do
-  echo "Clearing $folder" | sed 's/\/$//'
-  rm ${folder}main
-  echo -e "Done\n"
+  if [ -f ${folder}main ]; then
+    echo "Clearing $folder" | sed 's/\/$//'
+    rm ${folder}main
+    echo -e "Done\n"
+  fi
 done
